+++ assets/javascripts/kitten/components/layout/adaptable-grid/adaptable-grid.stories.js
@@ -1,34 +1,47 @@
 import React from 'react'
-import { text } from '@storybook/addon-knobs'
 import { AdaptableGrid, AdaptableGridCol } from './index'
-import { Container } from '../../../components/layout/container'
 
-export const Default = () => (
-  <Container>
-    <AdaptableGrid colNumber={10} as={text('Container tag', 'div')}>
-      <AdaptableGridCol
-        as={text('First child tag', 'div')}
-        col={4}
-        offset={3}
-        col-s={5}
-        offset-s={0}
-        col-l={2}
-      >
-        <div style={{ background: '#fae9d8', textAlign: 'center' }}>
-          {'🚲 '.repeat(5)}
-        </div>
-      </AdaptableGridCol>
-      <AdaptableGridCol
-        as={text('Second child tag', 'div')}
-        col-s={3}
-        offset-s={2}
-        col-l={6}
-        offset-l={1}
-      >
-        <div style={{ background: '#fae9d8', textAlign: 'center' }}>
-          {'🚲 '.repeat(8)}
-        </div>
-      </AdaptableGridCol>
-    </AdaptableGrid>
-  </Container>
+export const Default = args => (
+  <AdaptableGrid {...args}>
+    <AdaptableGridCol col={4} offset={3} col-s={5} offset-s={0} col-l={2}>
+      <div style={{ background: '#fae9d8', textAlign: 'center' }}>
+        {'🚲 '.repeat(5)}
+      </div>
+    </AdaptableGridCol>
+    <AdaptableGridCol col-s={3} offset-s={2} col-l={6} offset-l={1}>
+      <div style={{ background: '#fae9d8', textAlign: 'center' }}>
+        {'🚲 '.repeat(8)}
+      </div>
+    </AdaptableGridCol>
+  </AdaptableGrid>
 )
+
+Default.decorators = [
+  Story => (
+    <div className="story-Container">
+      <Story />
+    </div>
+  ),
+]
+
+Default.args = {
+  gutter: 20,
+  colNumber: 12,
+  colAlign: 'center',
+}
+
+Default.argTypes = {
+  gutter: {
+    name: 'gutter',
+    control: { type: 'number' },
+  },
+  colNumber: {
+    name: 'colNumber',
+    control: { type: 'number' },
+  },
+  colAlign: {
+    name: 'colAlign',
+    options: ['left', 'right', 'center'],
+    control: { type: 'select' },
+  },
+}