+++ assets/javascripts/kitten/components/layout/scrollable-container/stories.js
@@ -1,14 +1,25 @@
 import React from 'react'
 import { ScrollableContainer } from './index'
-import COLORS from '../../../constants/colors-config'
+import { COLORS } from '../../..'
 
 export default {
   title: 'Layout/ScrollableContainer',
   component: ScrollableContainer,
+  args: {
+    shadowColor: COLORS.background1,
+  },
+  argTypes: {
+    shadowColor: {
+      name: 'shadowColor',
+      control: {
+        type: 'color',
+      },
+    },
+  },
 }
 
-export const Default = () => (
-  <ScrollableContainer style={{ backgroundColor: COLORS.line1 }}>
+export const Default = args => (
+  <ScrollableContainer {...args} style={{ backgroundColor: COLORS.line1 }}>
     <div style={{ height: '100px', lineHeight: '100px' }}>
       {'🚗 '.repeat(100)}
     </div>