+++ assets/javascripts/kitten/components/layout/side-grid/stories.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { SideGrid, SideGridContent, SideGridAside } from './index'
-import { select } from '@storybook/addon-knobs'
 import {
   InformationBox,
   Marger,
@@ -13,13 +12,26 @@ import {
 export default {
   title: 'Layout/SideGrid',
   component: SideGrid,
+  args: {
+    asidePosition: 'end',
+    asideSize: 'default',
+  },
+  argTypes: {
+    asidePosition: {
+      name: 'asidePosition',
+      options: ['end', 'start'],
+      control: { type: 'select' },
+    },
+    asideSize: {
+      name: 'asideSize',
+      options: ['default', 'small', 'large'],
+      control: { type: 'select' },
+    },
+  },
 }
 
-export const Default = () => (
-  <SideGrid
-    asidePosition={select('asidePosition', ['end', 'start'], 'end')}
-    asideSize={select('asideSize', ['default', 'small', 'large'], 'default')}
-  >
+export const Default = args => (
+  <SideGrid {...args}>
     <SideGridContent>
       <Grid>
         <GridCol col-s="10" offset-s="1">